Initial revision of rit, the ruby hd remaster of git, from Buffalo
This commit records a minimal set of functionality necessary for the code to store itself as a valid Git commit. This includes writing the following object types to the database:
- Blobs of ASCII text
- Trees containing a flat list of regular files
- Commits that contain a tree pointer, author info and message
These objects are written to , compressed using zlib.
At this stage, there is no index and no  command; the  command simply writes everything in the working tree to the database and commits it.
--
Second commit of RIT, adds parent commit support

Third commit of RIT.

- This adds tree support.  Refactors project into /lib/ and /bin/ directories.